Module google_playmoviespartner1::api

Structs

AccountAvailGetCall
Get an Avail given its avail group id and avail id.

AccountAvailListCall
List Avails owned or managed by the partner.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the PlayMovies hub.

AccountOrderGetCall
Get an Order given its id.

AccountOrderListCall
List Orders owned or managed by the partner.

AccountStoreInfoCountryGetCall
Get a StoreInfo given its video id and country.

AccountStoreInfoListCall
List StoreInfos owned or managed by the partner.

Avail
An Avail describes the Availability Window of a specific Edit in a given country, which means the period Google is allowed to sell or rent the Edit.

ListAvailsResponse
Response to the 'ListAvails' method.

ListOrdersResponse
Response to the 'ListOrders' method.

ListStoreInfosResponse
Response to the 'ListStoreInfos' method.

Order
An Order tracks the fulfillment of an Edit when delivered using the legacy, non-component-based delivery.

PlayMovies
Central instance to access all PlayMovies related resource activities

StoreInfo
Information about a playable sequence (video) associated with an Edit and available at the Google Play Store.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.