Struct google_playmoviespartner1::AccountMethods

pub struct AccountMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on account resources. It is not used directly, but through the Playmoviespartner hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_playmoviespartner1 as playmoviespartner1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use playmoviespartner1::Playmoviespartner;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Playmoviespartner::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `avails_list(...)`, `experience_locales_get(...)`, `experience_locales_list(...)`, `orders_get(...)`, `orders_list(...)`, `store_infos_country_get(...)` and `store_infos_list(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, A> AccountMethods<'a, C, A>

fn orders_list(&self, account_id: &str) -> AccountOrderListCall<'a, C, A>

Create a builder to help you perform the following task:

List Orders owned or managed by the partner. See Authentication and Authorization rules and List methods rules for more information about this method.

Arguments

accountId - REQUIRED. See General rules for more information about this field.

fn store_infos_country_get(&self, account_id: &str, video_id: &str, country: &str) -> AccountStoreInfoCountryGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get a StoreInfo given its video id and country. See Authentication and Authorization rules and Get methods rules for more information about this method.

Arguments

accountId - REQUIRED. See General rules for more information about this field.

videoId - REQUIRED. Video ID.

country - REQUIRED. Edit country.

fn orders_get(&self, account_id: &str, order_id: &str) -> AccountOrderGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get an Order given its id. See Authentication and Authorization rules and Get methods rules for more information about this method.

Arguments

accountId - REQUIRED. See General rules for more information about this field.

orderId - REQUIRED. Order ID.

fn avails_list(&self, account_id: &str) -> AccountAvailListCall<'a, C, A>

Create a builder to help you perform the following task:

List Avails owned or managed by the partner. See Authentication and Authorization rules and List methods rules for more information about this method.

Arguments

accountId - REQUIRED. See General rules for more information about this field.

fn experience_locales_get(&self, account_id: &str, el_id: &str) -> AccountExperienceLocaleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get an ExperienceLocale given its id. See Authentication and Authorization rules and Get methods rules for more information about this method.

Arguments

accountId - REQUIRED. See General rules for more information about this field.

elId - REQUIRED. ExperienceLocale ID, as defined by Google.

fn store_infos_list(&self, account_id: &str) -> AccountStoreInfoListCall<'a, C, A>

Create a builder to help you perform the following task:

List StoreInfos owned or managed by the partner. See Authentication and Authorization rules and List methods rules for more information about this method.

Arguments

accountId - REQUIRED. See General rules for more information about this field.

fn experience_locales_list(&self, account_id: &str) -> AccountExperienceLocaleListCall<'a, C, A>

Create a builder to help you perform the following task:

List ExperienceLocales owned or managed by the partner. See Authentication and Authorization rules and List methods rules for more information about this method.

Arguments

accountId - REQUIRED. See General rules for more information about this field.