Struct google_playmoviespartner1::Avail

pub struct Avail {
    pub pph_names: Option<Vec<String>>,
    pub license_type: Option<String>,
    pub encode_id: Option<String>,
    pub caption_included: Option<bool>,
    pub release_date: Option<String>,
    pub episode_alt_id: Option<String>,
    pub episode_title_internal_alias: Option<String>,
    pub avail_id: Option<String>,
    pub end: Option<String>,
    pub alt_id: Option<String>,
    pub series_alt_id: Option<String>,
    pub start: Option<String>,
    pub title_internal_alias: Option<String>,
    pub season_title_internal_alias: Option<String>,
    pub territory: Option<String>,
    pub product_id: Option<String>,
    pub store_language: Option<String>,
    pub season_number: Option<String>,
    pub episode_number: Option<String>,
    pub format_profile: Option<String>,
    pub price_value: Option<String>,
    pub video_id: Option<String>,
    pub rating_reason: Option<String>,
    pub caption_exemption: Option<String>,
    pub work_type: Option<String>,
    pub display_name: Option<String>,
    pub season_alt_id: Option<String>,
    pub content_id: Option<String>,
    pub rating_system: Option<String>,
    pub rating_value: Option<String>,
    pub series_title_internal_alias: Option<String>,
    pub suppression_lift_date: Option<String>,
    pub price_type: Option<String>,
}

An Avail describes the Availability Window of a specific Edit in a given country, which means the period Google is allowed to sell or rent the Edit. Avails are exposed in EMA format Version 1.6b (available at http://www.movielabs.com/md/avails/) Studios can see the Avails for the Titles they own. Post-production houses cannot see any Avails.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

avails get accounts (response)

Fields
pph_names: Option<Vec<String>>

Name of the post-production houses that manage the Avail. Not part of EMA Specs.

license_type: Option<String>

Type of transaction.

encode_id: Option<String>

Manifestation Identifier. This should be the Manifestation Level EIDR. Example: "10.2340/1489-49A2-3956-4B2D-FE16-7"

caption_included: Option<bool>

Communicating if caption file will be delivered.

release_date: Option<String>

Release date of the Title in earliest released territory. Typically it is just the year, but it is free-form as per EMA spec. Examples: "1979", "Oct 2014"

episode_alt_id: Option<String>

Other identifier referring to the episode, as defined by partner. Only available on TV avails. Example: "rs_googlers_s1_3".

episode_title_internal_alias: Option<String>

OPTIONAL.TV Only. Title used by involved parties to refer to this episode. Only available on TV Avails. Example: "Coding at Google".

avail_id: Option<String>

ID internally generated by Google to uniquely identify an Avail. Not part of EMA Specs.

end: Option<String>

End of term in YYYY-MM-DD format in the timezone of the country of the Avail. "Open" if no end date is available. Example: "2019-02-17"

alt_id: Option<String>

Other identifier referring to the Edit, as defined by partner. Example: "GOOGLER_2006"

series_alt_id: Option<String>

Other identifier referring to the series, as defined by partner. Only available on TV avails. Example: "rs_googlers".

start: Option<String>

Start of term in YYYY-MM-DD format in the timezone of the country of the Avail. Example: "2013-05-14".

title_internal_alias: Option<String>

Title used by involved parties to refer to this content. Example: "Googlers, The". Only available on Movie Avails.

season_title_internal_alias: Option<String>

Title used by involved parties to refer to this season. Only available on TV Avails. Example: "Googlers, The".

territory: Option<String>

ISO 3166-1 alpha-2 country code for the country or territory of this Avail. For Avails, we use Territory in lieu of Country to comply with EMA specifications. But please note that Territory and Country identify the same thing. Example: "US".

product_id: Option<String>

Edit Identifier. This should be the Edit Level EIDR. Example: "10.2340/1489-49A2-3956-4B2D-FE16-6"

store_language: Option<String>

Spoken language of the intended audience. Language shall be encoded in accordance with RFC 5646. Example: "fr".

season_number: Option<String>

The number assigned to the season within a series. Only available on TV Avails. Example: "1".

episode_number: Option<String>

The number assigned to the episode within a season. Only available on TV Avails. Example: "3".

format_profile: Option<String>

Indicates the format profile covered by the transaction.

price_value: Option<String>

Value to be applied to the pricing type. Example: "4" or "2.99"

video_id: Option<String>

Google-generated ID identifying the video linked to this Avail, once delivered. Not part of EMA Specs. Example: 'gtry456_xc'

rating_reason: Option<String>

Value representing the rating reason. Rating reasons should be formatted as per EMA ratings spec and comma-separated for inclusion of multiple reasons. Example: "L, S, V"

caption_exemption: Option<String>

Communicating an exempt category as defined by FCC regulations. It is not required for non-US Avails. Example: "1"

work_type: Option<String>

Work type as enumerated in EMA.

display_name: Option<String>

The name of the studio that owns the Edit referred in the Avail. This is the equivalent of studio_name in other resources, but it follows the EMA nomenclature. Example: "Google Films".

season_alt_id: Option<String>

Other identifier referring to the season, as defined by partner. Only available on TV avails. Example: "rs_googlers_s1".

content_id: Option<String>

Title Identifier. This should be the Title Level EIDR. Example: "10.5240/1489-49A2-3956-4B2D-FE16-5".

rating_system: Option<String>

Rating system applied to the version of title within territory of Avail. Rating systems should be formatted as per EMA ratings spec Example: "MPAA"

rating_value: Option<String>

Value representing the rating. Ratings should be formatted as per http://www.movielabs.com/md/ratings/ Example: "PG"

series_title_internal_alias: Option<String>

Title used by involved parties to refer to this series. Only available on TV Avails. Example: "Googlers, The".

suppression_lift_date: Option<String>

First date an Edit could be publically announced as becoming available at a specific future date in territory of Avail. Not the Avail start date or pre-order start date. Format is YYYY-MM-DD. Only available for pre-orders. Example: "2012-12-10"

price_type: Option<String>

Type of pricing that should be applied to this Avail based on how the partner classify them. Example: "Tier", "WSP", "SRP", or "Category".