Struct google_playmoviespartner1::Component

pub struct Component {
    pub status: Option<String>,
    pub alt_cut_ids: Option<Vec<String>>,
    pub processing_errors: Option<Vec<String>>,
    pub studio_name: Option<String>,
    pub status_detail: Option<String>,
    pub rejection_note: Option<String>,
    pub approved_time: Option<String>,
    pub title_level_eidrs: Option<Vec<String>>,
    pub el_ids: Option<Vec<String>>,
    pub name: Option<String>,
    pub language: Option<String>,
    pub playable_unit_type: Option<String>,
    pub component_detail_type: Option<String>,
    pub received_time: Option<String>,
    pub custom_ids: Option<Vec<String>>,
    pub pph_name: Option<String>,
    pub filename: Option<String>,
    pub priority: Option<f64>,
    pub edit_level_eidrs: Option<Vec<String>>,
    pub normalized_priority: Option<String>,
    pub type_: Option<String>,
    pub component_id: Option<String>,
}

A Component is an element (audio, video, subtitle, artwork, trailer, etc...) that is used in a set of ExperienceLocales. A Component is owned by a Studio and managed either by the Studio itself or by one Post-Production House. Each Component is identified by a component_id and its type.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

components type get accounts (response)

Fields
status: Option<String>

High-level status of the Component.

alt_cut_ids: Option<Vec<String>>

List of Alternative Cut IDs, sometimes available in lieu of the main Edit-level EIDR IDs. This is not an EIDR ID, but a Partner-provided ID. Example: "206346_79838".

processing_errors: Option<Vec<String>>

Processing errors during XML file parsing. Example: 'Invalid input file'

studio_name: Option<String>

Name of the studio that owns the Component.

status_detail: Option<String>

Detailed status of the Component

rejection_note: Option<String>

Notes explaining why a Component has been rejected.

approved_time: Option<String>

Timestamp when the Component was approved.

title_level_eidrs: Option<Vec<String>>

List of Title-level EIDR IDs. Example: "10.5240/1489-49A2-3956-4B2D-FE16-5".

el_ids: Option<Vec<String>>

IDs internally generated by Google to uniquely identify the ExperienceLocales for which the Component is used. Example: 'KRZiVjY9h7t'

name: Option<String>

Default Edit name, usually in the language of the country of origin. Example: "Googlers, The".

language: Option<String>

Language of the component, using the "BCP 47" format. Examples: "en", "en-US", "es", "es-419".

playable_unit_type: Option<String>

Type of the playable unit for which the Component is intended.

component_detail_type: Option<String>

Detail about the type of the Component.

received_time: Option<String>

Timestamp when the Component was received.

custom_ids: Option<Vec<String>>

List of custom IDs (defined by the partner) linked to the ExperienceLocale using this Component. Example: "R86241"

pph_name: Option<String>

Name of the post-production house that manages the Component.

filename: Option<String>

File name of the Component when delivered.

priority: Option<f64>

Component priority, as defined by Google. The higher the value, the higher the priority.

edit_level_eidrs: Option<Vec<String>>

List of Edit-level EIDR IDs. Example: "10.5240/1489-49A2-3956-4B2D-FE16-6".

normalized_priority: Option<String>

A simpler representation of the priority.

type_: Option<String>

Type of the Component. Example: AUDIO_51

component_id: Option<String>

ID internally generated by Google to uniquely identify the Component. Example: 'wteyrc_647xc'