Struct google_playmoviespartner1::ExperienceLocale

pub struct ExperienceLocale {
    pub pph_names: Option<Vec<String>>,
    pub status: Option<String>,
    pub title_level_eidr: Option<String>,
    pub trailer_id: Option<String>,
    pub channel_id: Option<String>,
    pub studio_name: Option<String>,
    pub video_id: Option<String>,
    pub approved_time: Option<String>,
    pub edit_level_eidr: Option<String>,
    pub inventory_id: Option<String>,
    pub playable_sequence_id: Option<String>,
    pub name: Option<String>,
    pub el_id: Option<String>,
    pub presentation_id: Option<String>,
    pub earliest_avail_start_time: Option<String>,
    pub language: Option<String>,
    pub alt_cut_id: Option<String>,
    pub custom_ids: Option<Vec<String>>,
    pub priority: Option<f64>,
    pub normalized_priority: Option<String>,
    pub created_time: Option<String>,
    pub country: Option<String>,
    pub type_: Option<String>,
}

An ExperienceLocale tracks the fulfillment of a Title in a country using a specific language, when delivered using component-based delivery. For example, a Title in Switzerland might have 3 ExperienceLocales: they both share the same country ("CH"), but each has different languages ("de", "fr", and "it"). Each ExperienceLocale is uniquely identified by an el_id, which is generated by Google. Externally, an ExperienceLocale can also be identified by partners using its EIDR IDs, AltCutID or custom_id (when provided).

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

experience locales get accounts (response)

Fields

pph_names

Name of the post-production houses that manage the ExperienceLocale.

status

High-level status of the ExperienceLocale.

title_level_eidr

Title-level EIDR ID. Example: "10.5240/1489-49A2-3956-4B2D-FE16-5".

trailer_id

Trailer ID, as defined by Google, linked to the trailer video in the ExperienceLocale. Example: 'gtry457_tr'.

channel_id

YouTube Channel ID linked to the ExperienceLocale. Example: "UCRG64darCZhb".

studio_name

Name of the studio that owns the ExperienceLocale.

video_id

Video ID, as defined by Google, linked to the feature video in the ExperienceLocale. Example: 'gtry456_xc'.

approved_time

Timestamp when the ExperienceLocale was approved.

edit_level_eidr

Edit-level EIDR ID. Example: "10.5240/1489-49A2-3956-4B2D-FE16-6".

inventory_id

InventoryID as defined in the EMA specs.

playable_sequence_id

PlayableSequenceID as defined in the EMA specs.

name

Default Edit name, usually in the language of the country of origin. Example: "Googlers, The".

el_id

ID internally generated by Google to uniquely identify a ExperienceLocale. Example: 'KRZiVjY9h7t'

presentation_id

PresentationID as defined in the EMA specs.

earliest_avail_start_time

Timestamp of the earliest start date of the Avails linked to this ExperienceLocale.

language

Language of the ExperienceLocale, using the "BCP 47" format. Examples: "en", "en-US", "es", "es-419".

alt_cut_id

Alternative Cut ID, sometimes available in lieu of the main Edit-level EIDR ID. This is not an EIDR ID, but a Partner-provided ID. Example: "206346_79838".

custom_ids

List of custom IDs (defined by the partner) linked to this ExperienceLocale. Example: "R86241"

priority

ExperienceLocale priority, as defined by Google. The higher the value, the higher the priority. Example: 90

normalized_priority

A simpler representation of the priority.

created_time

Timestamp when the ExperienceLocale was created.

country

Country where the ExperienceLocale is available, using the "ISO 3166-1 alpha-2" format. Example: "US".

type_

Type of the Edit linked to the ExperienceLocale.