Struct google_playmoviespartner1::StoreInfo

pub struct StoreInfo {
    pub pph_names: Option<Vec<String>>,
    pub has_est_offer: Option<bool>,
    pub title_level_eidr: Option<String>,
    pub has_sd_offer: Option<bool>,
    pub season_number: Option<String>,
    pub live_time: Option<String>,
    pub trailer_id: Option<String>,
    pub studio_name: Option<String>,
    pub video_id: Option<String>,
    pub has_hd_offer: Option<bool>,
    pub has_vod_offer: Option<bool>,
    pub edit_level_eidr: Option<String>,
    pub subtitles: Option<Vec<String>>,
    pub episode_number: Option<String>,
    pub show_id: Option<String>,
    pub name: Option<String>,
    pub country: Option<String>,
    pub season_id: Option<String>,
    pub mid: Option<String>,
    pub has_audio51: Option<bool>,
    pub season_name: Option<String>,
    pub audio_tracks: Option<Vec<String>>,
    pub has_info_cards: Option<bool>,
    pub type_: Option<String>,
    pub show_name: Option<String>,
}

Information about a playable sequence (video) associated with an Edit and available at the Google Play Store. Internally, each StoreInfo is uniquely identified by a video_id and country. Externally, Title-level EIDR or Edit-level EIDR, if provided, can also be used to identify a specific title or edit in a country.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

store infos country get accounts (response)

Fields
pph_names: Option<Vec<String>>

Name of the post-production houses that manage the Edit.

has_est_offer: Option<bool>

Whether the Edit has a EST offer.

title_level_eidr: Option<String>

Title-level EIDR ID. Example: "10.5240/1489-49A2-3956-4B2D-FE16-5".

has_sd_offer: Option<bool>

Whether the Edit has a SD offer.

season_number: Option<String>

The number assigned to the season within a show. Only available on TV Edits. Example: "1".

live_time: Option<String>

Timestamp when the Edit went live on the Store.

trailer_id: Option<String>

Google-generated ID identifying the trailer linked to the Edit. Example: 'bhd_4e_cx'

studio_name: Option<String>

Name of the studio that owns the Edit ordered.

video_id: Option<String>

Google-generated ID identifying the video linked to the Edit. Example: 'gtry456_xc'

has_hd_offer: Option<bool>

Whether the Edit has a HD offer.

has_vod_offer: Option<bool>

Whether the Edit has a VOD offer.

edit_level_eidr: Option<String>

Edit-level EIDR ID. Example: "10.5240/1489-49A2-3956-4B2D-FE16-6".

subtitles: Option<Vec<String>>

Subtitles available for this Edit.

episode_number: Option<String>

The number assigned to the episode within a season. Only available on TV Edits. Example: "1".

show_id: Option<String>

Google-generated ID identifying the show linked to the Edit. Only available for TV Edits. Example: 'et2hsue_x'

name: Option<String>

Default Edit name, usually in the language of the country of origin. Example: "Googlers, The".

country: Option<String>

Country where Edit is available in ISO 3166-1 alpha-2 country code. Example: "US".

season_id: Option<String>

Google-generated ID identifying the season linked to the Edit. Only available for TV Edits. Example: 'ster23ex'

mid: Option<String>

Knowledge Graph ID associated to this Edit, if available. This ID links the Edit to its knowledge entity, externally accessible at http://freebase.com. In the absense of Title EIDR or Edit EIDR, this ID helps link together multiple Edits across countries. Example: '/m/0ffx29'

has_audio51: Option<bool>

Whether the Edit has a 5.1 channel audio track.

season_name: Option<String>

Default Season name, usually in the language of the country of origin. Only available for TV Edits Example: "Googlers, The - A Brave New World".

audio_tracks: Option<Vec<String>>

Audio tracks available for this Edit.

has_info_cards: Option<bool>

Whether the Edit has info cards.

type_: Option<String>

Edit type, like Movie, Episode or Season.

show_name: Option<String>

Default Show name, usually in the language of the country of origin. Only available for TV Edits Example: "Googlers, The".