Struct google_playmoviespartner1::Order

pub struct Order {
    pub order_id: Option<String>,
    pub status: Option<String>,
    pub channel_id: Option<String>,
    pub studio_name: Option<String>,
    pub status_detail: Option<String>,
    pub rejection_note: Option<String>,
    pub approved_time: Option<String>,
    pub video_id: Option<String>,
    pub episode_name: Option<String>,
    pub name: Option<String>,
    pub season_name: Option<String>,
    pub earliest_avail_start_time: Option<String>,
    pub countries: Option<Vec<String>>,
    pub normalized_priority: Option<String>,
    pub received_time: Option<String>,
    pub ordered_time: Option<String>,
    pub pph_name: Option<String>,
    pub priority: Option<f64>,
    pub legacy_priority: Option<String>,
    pub custom_id: Option<String>,
    pub channel_name: Option<String>,
    pub type_: Option<String>,
    pub show_name: Option<String>,
}

An Order tracks the fulfillment of an Edit when delivered using the legacy, non-component-based delivery. Each Order is uniquely identified by an order_id, which is generated by Google. Externally, Orders can also be identified by partners using its custom_id (when provided).

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

orders get accounts (response)

Fields
order_id: Option<String>

ID internally generated by Google to uniquely identify an Order. Example: 'abcde12_x'

status: Option<String>

High-level status of the order.

channel_id: Option<String>

YouTube Channel ID that should be used to fulfill the Order. Example: "UCRG64darCZhb".

studio_name: Option<String>

Name of the studio that owns the Edit ordered.

status_detail: Option<String>

Detailed status of the order

rejection_note: Option<String>

Field explaining why an Order has been rejected. Example: "Trailer audio is 2ch mono, please re-deliver in stereo".

approved_time: Option<String>

Timestamp when the Order was approved.

video_id: Option<String>

Google-generated ID identifying the video linked to this Order, once delivered. Example: 'gtry456_xc'.

episode_name: Option<String>

Default Episode name, usually in the language of the country of origin. Only available for TV Edits Example: "Googlers, The - Pilot".

name: Option<String>

Default Edit name, usually in the language of the country of origin. Example: "Googlers, The".

season_name: Option<String>

Default Season name, usually in the language of the country of origin. Only available for TV Edits Example: "Googlers, The - A Brave New World".

earliest_avail_start_time: Option<String>

Timestamp of the earliest start date of the Avails linked to this Order.

countries: Option<Vec<String>>

Countries where the Order is available, using the "ISO 3166-1 alpha-2" format (example: "US").

normalized_priority: Option<String>

A simpler representation of the priority.

received_time: Option<String>

Timestamp when the Order was fulfilled.

ordered_time: Option<String>

Timestamp when the Order was created.

pph_name: Option<String>

Name of the post-production house that manages the Edit ordered.

priority: Option<f64>

Order priority, as defined by Google. The higher the value, the higher the priority. Example: 90

legacy_priority: Option<String>

Legacy Order priority, as defined by Google. Example: 'P0'

custom_id: Option<String>

ID that can be used to externally identify an Order. This ID is provided by partners when submitting the Avails. Example: 'GOOGLER_2006'

channel_name: Option<String>

YouTube Channel Name that should be used to fulfill the Order. Example: "Google_channel".

type_: Option<String>

Type of the Edit linked to the Order.

show_name: Option<String>

Default Show name, usually in the language of the country of origin. Only available for TV Edits Example: "Googlers, The".